requirements.txt
azure-identity
azure-keyvault-keys
pyyaml